wip: added tagsmodal,output and detailsoutput on projectspage
